pyproject.toml
[tool.poetry]
name = "pyvidia-update"
version = "0.1.0"
description = ""
authors = ["Maalik Kaufhold <mail@maalikaufhold.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
ruff-lsp = "^0.0.54"
pre-commit = "^3.7.1"


[tool.poetry.group.scraper.dependencies]
selenium = "^4.22.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
